requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy
pillow
opencv-python
matplotlib
pycocotools
tqdm
transformers
einops
segment-anything-2
jupyterlab
wandb
accelerate